perf audit: narrow node handles, break the val_count recursion, borrow keys in next_items
docs/PERF_AUDIT_2026-09-17.md is the audit (JET @report_opt / @report_call, Aqua, Profile.Allocs,
AllocCheck, BenchmarkTools) with the probes under ~/csai-work/gates/audit/.

Fixed:
- src/nodes/NodeVariant.jl (new): the closed node union `TrieNodeVariant`, `_fnode`, and BOTH `as_tagged`
  methods, moved out of TrieNode.jl because the union can only be written after the last node type. The
  accessor now ASSERTS into the closed union, so calls through it union-split instead of dispatching;
  seven write-zipper `.node` reads go through it too. Runtime-dispatch sites over nine entry points
  96 -> 62 (remove_val_at! 15 -> 6, join_map_into! 32 -> 20, meet_into! 35 -> 22).
- `node_val_count` / `val_count_below_node` / `val_count_below_root` annotated `::Int`: they are mutually
  recursive across six node types, inference abandoned the cycle and returned `Any`, which propagated out
  through `val_count`. Public API inferring `Any` (real cases): 4 -> 0.
- `next_items` hands back a BORROWED `NodeKeySlice` instead of a fresh Vector per step, a byte node slicing
  the static ALL_BYTES table as upstream does (dense_byte_node.rs:1052). 5 000-value walk:
  83 901 -> 72 898 allocations; AllocCheck sites for to_next_val! 42 -> 37.

🔴 One regression this audit CAUSED and caught: narrowing the read at `remove_val!` cost 18 fuzz cases,
because `node_remove_val!(::Nothing)` returns nothing while `(::EmptyNode)` ERRORS. I had checked that the
EmptyNode method existed without reading its body. Reverted with a comment; the other six sites narrow
safely because both of their methods agree.

Measured and NOT adopted (both made allocations worse, 72 898 -> 83 901): guarding the observer slices,
and `no_key()` in to_next_get_val_observed!'s exhausted branch.

⚠️ Timings on this box vary 3-4x run to run, so every claim rests on allocation counts and static dispatch
sites, which are deterministic. Aqua is clean. GC time was 0 ms in every workload.

Warm: PathMap 172/172 (Lean 2 known, fuzz 2995/3000, curated 46/46, AllocCheck pins), MORK 8171/8171.

+++ src/nodes/TinyRefNode.jl
@@ -174,7 +174,7 @@ iter_token_for_path(::TinyRefNode, ::AbstractVector{UInt8}) =
 ascend_iter_token(::TinyRefNode, ::IterToken, ::Int) = error("TinyRefNode::ascend_iter_token — unreachable")
 next_items(::TinyRefNode, ::IterToken, ::Bool) = error("TinyRefNode::next_items — unreachable")
 
-function node_val_count(t::TinyRefNode, cache::Dict{UInt64, Int})
+function node_val_count(t::TinyRefNode, cache::Dict{UInt64, Int})::Int
     node_val_count(into_full(t), cache)
 end
